fix(expired): remove items from elastic first
there are items in elastic without an item in mongo,
so rather remove it first from elastic and if that works
remove it from mongo, otherwise there is something showing up
that shouldn't be there and user actions fail on it.

plus remove expired ingest items from elastic which shouldn't be there
to clean up the mess.

SD-4793