[publishing] encoding handling in formatter and transmitter
formatter can now specify the final encoding using "item_encoding" encoding, to do this they can return a list of dictionnaries instead of a list of tuples.
Only file_output transmitter was testing if "formatted_item" is a bytes or str, so this part was removed, and it is now assumed that "formatted_item" is always str, while "encoded_item" will be bytes.
If a formatter want to return bytes, it can used "encoded_item", but "formatted_item" still need to be filled with either a unicode friendly version, or an empty string. It is used at least for preview.

needed for SDNTB-253